+++ routing-bird-health/src/main/java/com/jivesoftware/os/routing/bird/health/checkers/DiskFreeHealthChecker.java
@@ -41,7 +41,7 @@ public class DiskFreeHealthChecker extends MinMaxHealthChecker implements Schedu
                 }
             }
             Counter counter = new Counter(ValueType.RATE);
-            counter.set((int) (worstHealth * 100));
+            counter.set((long) (worstHealth * 100));
             check(counter, sb.toString(), "Remove some unused files or get more or larger drive.");
         } catch (Exception x) {
             // TODO what?
